requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cleanup==6.0.0
mysql-connector-python==8.0.28
mysqlclient==2.1.0
Pillow==9.1.0
protobuf==3.20.0
PyMySQL==1.0.2
sqlparse==0.4.2
tzdata==2022.1
